Place custom image as avatar in header
The forced use of Gravatar for the avatar is a problem because it
requires creating an account on Gravatar. To get a special displayed
image there, it implies creating a fake account and fake email.

This solution is still not good, as the image is hard coded. This must
be changed later.

Signed-off-by: Cilyan Olowen <gaknar@gmail.com>