Move to SCSS and Font-Awesome
Leaving behind Genericons which is too poor at the moment, especially
converning brands available for social links, the style was fitted to
Font-Awesome.

The integration was done with the SCSS version of Font-Awesome to
include only the necessary parts and keep the style small and clean.

This implied spliting the actual style into SCSS fragments, but it's a
rough CSS copy at the moment, so the code is not very nice yet.

Signed-off-by: Cilyan Olowen <gaknar@gmail.com>