Add Rake task to package release
Add a task to complete the release process: package the result in a nice
zip file, ready to be imported in Wordpress.

The task uses the shell to run `7z`. This is not ideal, but `PackageTask`
refuses to include the proper files, whatever the configuration I give
at the moment. I hope to understand the configuration better later. At
least it now works, and 7-zip is available on all platforms.

Signed-off-by: Cilyan Olowen <gaknar@gmail.com>