Support for local avatar
Instead of using a Gravatar account, one can now choose a locally
uploaded image. The image is expected to be 70x70px. If no images is
selected (or the selection is removed), the Gravatar image is used like
in the original Fictive theme (i.e. the admin email is used except if
another email is given in the theme customisation page). If both image
is not selected and gravatar email is set to empty, then the avatar will
not be displayed.

Localisation update required.

Signed-off-by: Cilyan Olowen <gaknar@gmail.com>